XP-4271 Unpublish dialog - Add progress bar
Moved PublishDialog dependencies to external file.
