+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-studio/js/app/dialog/ProgressBarDialog.ts
@@ -2,8 +2,6 @@ import "../../api.ts";
 import {DependantItemsDialog} from "../dialog/DependantItemsDialog";
 import {ContentPublishMenuManager} from "../browse/ContentPublishMenuManager";
 
-import ContentSummaryAndCompareStatus = api.content.ContentSummaryAndCompareStatus;
-
 export class ProcessingStats {
     // If the content is still being published after this time, show the progress bar (in ms)
     static progressBarDelay: number = 100; // 2000
@@ -17,10 +15,6 @@ export class ProcessingStats {
 
 export class ProgressBarDialog extends DependantItemsDialog {
 
-    // stashes previous checkbox state items, until selected items changed
-    // private stash: {[checked: string]: ContentSummaryAndCompareStatus[]} = {};
-    // private stashedCount: {[checked: string]: number} = {};
-    //
     private progressBar: api.ui.ProgressBar;
 
     constructor(dialogName: string, dialogSubName: string, dependantsName: string) {
@@ -82,50 +76,42 @@ export class ProgressBarDialog extends DependantItemsDialog {
         this.hide();
     }
 
-    //
-    // private pollPublishTask(taskId: api.task.TaskId, elapsed: number = 0, interval: number = ProcessingStats.pollInterval) {
-    //     setTimeout(() => {
-    //         if (!this.isProgressBarEnabled() && elapsed >= ProcessingStats.progressBarDelay) {
-    //             this.enableProgressBar();
-    //         }
-    //
-    //         new api.task.GetTaskInfoRequest(taskId).sendAndParse().then((task: api.task.TaskInfo) => {
-    //             let state = task.getState();
-    //             if (!task) {
-    //                 return; // task probably expired, stop polling
-    //             }
-    //
-    //             let progress = task.getProgress();
-    //
-    //             if (state == api.task.TaskState.FINISHED) {
-    //                 this.setProgressValue(100);
-    //                 this.onPublishComplete();
-    //
-    //                 api.notify.showSuccess(progress.getInfo());
-    //             } else if (state == api.task.TaskState.FAILED) {
-    //                 this.onPublishComplete();
-    //
-    //                 api.notify.showError('Publishing failed: ' + progress.getInfo());
-    //             } else {
-    //                 this.setProgressValue(task.getProgressPercentage());
-    //                 this.pollPublishTask(taskId, elapsed + interval, interval);
-    //             }
-    //
-    //         }).catch((reason: any) => {
-    //             this.onPublishComplete();
-    //
-    //             api.DefaultErrorHandler.handle(reason);
-    //         }).done();
-    //
-    //     }, interval);
-    // }
-    //
-    // private setProgressValue(value: number) {
-    //     if (this.isProgressBarEnabled()) {
-    //         this.progressBar.setValue(value);
-    //         if (!api.dom.Body.get().isShowingModalDialog()) {
-    //             ContentPublishMenuManager.getProgressBar().setValue(value);
-    //         }
-    //     }
-    // }
+
+    protected pollPublishTask(taskId: api.task.TaskId, elapsed: number = 0) {
+        const interval = ProcessingStats.pollInterval;
+        setTimeout(() => {
+            if (!this.isProgressBarEnabled() && elapsed >= ProcessingStats.progressBarDelay) {
+                this.enableProgressBar();
+            }
+
+            new api.task.GetTaskInfoRequest(taskId).sendAndParse().then((task: api.task.TaskInfo) => {
+                let state = task.getState();
+                if (!task) {
+                    return; // task probably expired, stop polling
+                }
+
+                let progress = task.getProgress();
+
+                if (state == api.task.TaskState.FINISHED) {
+                    this.setProgressValue(100);
+                    this.onPublishComplete();
+
+                    api.notify.showSuccess(progress.getInfo());
+                } else if (state == api.task.TaskState.FAILED) {
+                    this.onPublishComplete();
+
+                    api.notify.showError('Publishing failed: ' + progress.getInfo());
+                } else {
+                    this.setProgressValue(task.getProgressPercentage());
+                    this.pollPublishTask(taskId, elapsed + interval);
+                }
+
+            }).catch((reason: any) => {
+                this.onPublishComplete();
+
+                api.DefaultErrorHandler.handle(reason);
+            }).done();
+
+        }, interval);
+    }
 }
